@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.21.2
+// - protoc-gen-go-grpc v1.3.0
+// - protoc             v3.21.10
 // source: proto/v1/helloworld.proto
 
 package v1
@@ -18,12 +18,23 @@ import (
 // Requires gRPC-Go v1.32.0 or later.
 const _ = grpc.SupportPackageIsVersion7
 
+const (
+	HelloworldService_Login_FullMethodName      = "/helloworld.api.v1.HelloworldService/Login"
+	HelloworldService_Logout_FullMethodName     = "/helloworld.api.v1.HelloworldService/Logout"
+	HelloworldService_User_FullMethodName       = "/helloworld.api.v1.HelloworldService/User"
+	HelloworldService_UpdateInfo_FullMethodName = "/helloworld.api.v1.HelloworldService/UpdateInfo"
+	HelloworldService_Notify_FullMethodName     = "/helloworld.api.v1.HelloworldService/Notify"
+	HelloworldService_SayHello_FullMethodName   = "/helloworld.api.v1.HelloworldService/SayHello"
+)
+
 // HelloworldServiceClient is the client API for HelloworldService service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type HelloworldServiceClient interface {
 	// user service
 	Login(ctx context.Context, in *LoginRequest, opts ...grpc.CallOption) (*LoginResponse, error)
+	// 登出接口
+	Logout(ctx context.Context, in *LogoutRequest, opts ...grpc.CallOption) (*LogoutResponse, error)
 	User(ctx context.Context, in *UserRequest, opts ...grpc.CallOption) (*UserInfo, error)
 	UpdateInfo(ctx context.Context, in *UpdateInfoRequest, opts ...grpc.CallOption) (*UpdateInfoResponse, error)
 	Notify(ctx context.Context, in *NotifyRequest, opts ...grpc.CallOption) (*NotifyResponse, error)
@@ -41,7 +52,16 @@ func NewHelloworldServiceClient(cc grpc.ClientConnInterface) HelloworldServiceCl
 
 func (c *helloworldServiceClient) Login(ctx context.Context, in *LoginRequest, opts ...grpc.CallOption) (*LoginResponse, error) {
 	out := new(LoginResponse)
-	err := c.cc.Invoke(ctx, "/helloworld.api.v1.HelloworldService/Login", in, out, opts...)
+	err := c.cc.Invoke(ctx, HelloworldService_Login_FullMethodName, in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *helloworldServiceClient) Logout(ctx context.Context, in *LogoutRequest, opts ...grpc.CallOption) (*LogoutResponse, error) {
+	out := new(LogoutResponse)
+	err := c.cc.Invoke(ctx, HelloworldService_Logout_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -50,7 +70,7 @@ func (c *helloworldServiceClient) Login(ctx context.Context, in *LoginRequest, o
 
 func (c *helloworldServiceClient) User(ctx context.Context, in *UserRequest, opts ...grpc.CallOption) (*UserInfo, error) {
 	out := new(UserInfo)
-	err := c.cc.Invoke(ctx, "/helloworld.api.v1.HelloworldService/User", in, out, opts...)
+	err := c.cc.Invoke(ctx, HelloworldService_User_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -59,7 +79,7 @@ func (c *helloworldServiceClient) User(ctx context.Context, in *UserRequest, opt
 
 func (c *helloworldServiceClient) UpdateInfo(ctx context.Context, in *UpdateInfoRequest, opts ...grpc.CallOption) (*UpdateInfoResponse, error) {
 	out := new(UpdateInfoResponse)
-	err := c.cc.Invoke(ctx, "/helloworld.api.v1.HelloworldService/UpdateInfo", in, out, opts...)
+	err := c.cc.Invoke(ctx, HelloworldService_UpdateInfo_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -68,7 +88,7 @@ func (c *helloworldServiceClient) UpdateInfo(ctx context.Context, in *UpdateInfo
 
 func (c *helloworldServiceClient) Notify(ctx context.Context, in *NotifyRequest, opts ...grpc.CallOption) (*NotifyResponse, error) {
 	out := new(NotifyResponse)
-	err := c.cc.Invoke(ctx, "/helloworld.api.v1.HelloworldService/Notify", in, out, opts...)
+	err := c.cc.Invoke(ctx, HelloworldService_Notify_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -77,7 +97,7 @@ func (c *helloworldServiceClient) Notify(ctx context.Context, in *NotifyRequest,
 
 func (c *helloworldServiceClient) SayHello(ctx context.Context, in *HelloworldRequset, opts ...grpc.CallOption) (*HelloworldReply, error) {
 	out := new(HelloworldReply)
-	err := c.cc.Invoke(ctx, "/helloworld.api.v1.HelloworldService/SayHello", in, out, opts...)
+	err := c.cc.Invoke(ctx, HelloworldService_SayHello_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -90,6 +110,8 @@ func (c *helloworldServiceClient) SayHello(ctx context.Context, in *HelloworldRe
 type HelloworldServiceServer interface {
 	// user service
 	Login(context.Context, *LoginRequest) (*LoginResponse, error)
+	// 登出接口
+	Logout(context.Context, *LogoutRequest) (*LogoutResponse, error)
 	User(context.Context, *UserRequest) (*UserInfo, error)
 	UpdateInfo(context.Context, *UpdateInfoRequest) (*UpdateInfoResponse, error)
 	Notify(context.Context, *NotifyRequest) (*NotifyResponse, error)
@@ -104,6 +126,9 @@ type UnimplementedHelloworldServiceServer struct {
 
 func (UnimplementedHelloworldServiceServer) Login(context.Context, *LoginRequest) (*LoginResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Login not implemented")
+}
+func (UnimplementedHelloworldServiceServer) Logout(context.Context, *LogoutRequest) (*LogoutResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method Logout not implemented")
 }
 func (UnimplementedHelloworldServiceServer) User(context.Context, *UserRequest) (*UserInfo, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method User not implemented")
@@ -140,10 +165,28 @@ func _HelloworldService_Login_Handler(srv interface{}, ctx context.Context, dec 
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/helloworld.api.v1.HelloworldService/Login",
+		FullMethod: HelloworldService_Login_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(HelloworldServiceServer).Login(ctx, req.(*LoginRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _HelloworldService_Logout_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(LogoutRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(HelloworldServiceServer).Logout(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: HelloworldService_Logout_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(HelloworldServiceServer).Logout(ctx, req.(*LogoutRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -158,7 +201,7 @@ func _HelloworldService_User_Handler(srv interface{}, ctx context.Context, dec f
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/helloworld.api.v1.HelloworldService/User",
+		FullMethod: HelloworldService_User_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(HelloworldServiceServer).User(ctx, req.(*UserRequest))
@@ -176,7 +219,7 @@ func _HelloworldService_UpdateInfo_Handler(srv interface{}, ctx context.Context,
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/helloworld.api.v1.HelloworldService/UpdateInfo",
+		FullMethod: HelloworldService_UpdateInfo_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(HelloworldServiceServer).UpdateInfo(ctx, req.(*UpdateInfoRequest))
@@ -194,7 +237,7 @@ func _HelloworldService_Notify_Handler(srv interface{}, ctx context.Context, dec
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/helloworld.api.v1.HelloworldService/Notify",
+		FullMethod: HelloworldService_Notify_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(HelloworldServiceServer).Notify(ctx, req.(*NotifyRequest))
@@ -212,7 +255,7 @@ func _HelloworldService_SayHello_Handler(srv interface{}, ctx context.Context, d
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/helloworld.api.v1.HelloworldService/SayHello",
+		FullMethod: HelloworldService_SayHello_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(HelloworldServiceServer).SayHello(ctx, req.(*HelloworldRequset))
@@ -230,6 +273,10 @@ var HelloworldService_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "Login",
 			Handler:    _HelloworldService_Login_Handler,
+		},
+		{
+			MethodName: "Logout",
+			Handler:    _HelloworldService_Logout_Handler,
 		},
 		{
 			MethodName: "User",
