@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
 // 	protoc-gen-go v1.28.1
-// 	protoc        v3.21.2
+// 	protoc        v3.21.10
 // source: proto/v1/message.user.proto
 
 package v1
@@ -187,13 +187,16 @@ func (x *UserInfo) GetGender() UserInfo_Gender {
 	return UserInfo_SECRET
 }
 
+// 登录请求参数 示例是使用自定义tag 可以使用以下写法定义自己的tag
 type LoginRequest struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	UserName *string `protobuf:"bytes,1,opt,name=userName,proto3,oneof" json:"userName,omitempty"` // 用户名
-	Password *string `protobuf:"bytes,2,opt,name=password,proto3,oneof" json:"password,omitempty"` // 密码
+	// @gotags: form:"userName" json:"user_name" binding:"required"
+	UserName *string `protobuf:"bytes,1,opt,name=userName,proto3,oneof" json:"user_name" form:"userName" binding:"required"` // 用户名
+	// @gotags: form:"password" json:"password" binding:"required"
+	Password *string `protobuf:"bytes,2,opt,name=password,proto3,oneof" json:"password" form:"password" binding:"required"` // 密码
 }
 
 func (x *LoginRequest) Reset() {
@@ -289,6 +292,82 @@ func (x *LoginResponse) GetLoginedAt() *timestamppb.Timestamp {
 	return nil
 }
 
+type LogoutRequest struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+}
+
+func (x *LogoutRequest) Reset() {
+	*x = LogoutRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_proto_v1_message_user_proto_msgTypes[4]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *LogoutRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*LogoutRequest) ProtoMessage() {}
+
+func (x *LogoutRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_proto_v1_message_user_proto_msgTypes[4]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use LogoutRequest.ProtoReflect.Descriptor instead.
+func (*LogoutRequest) Descriptor() ([]byte, []int) {
+	return file_proto_v1_message_user_proto_rawDescGZIP(), []int{4}
+}
+
+type LogoutResponse struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+}
+
+func (x *LogoutResponse) Reset() {
+	*x = LogoutResponse{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_proto_v1_message_user_proto_msgTypes[5]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *LogoutResponse) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*LogoutResponse) ProtoMessage() {}
+
+func (x *LogoutResponse) ProtoReflect() protoreflect.Message {
+	mi := &file_proto_v1_message_user_proto_msgTypes[5]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use LogoutResponse.ProtoReflect.Descriptor instead.
+func (*LogoutResponse) Descriptor() ([]byte, []int) {
+	return file_proto_v1_message_user_proto_rawDescGZIP(), []int{5}
+}
+
 type UpdateInfoRequest struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
@@ -300,7 +379,7 @@ type UpdateInfoRequest struct {
 func (x *UpdateInfoRequest) Reset() {
 	*x = UpdateInfoRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_v1_message_user_proto_msgTypes[4]
+		mi := &file_proto_v1_message_user_proto_msgTypes[6]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -313,7 +392,7 @@ func (x *UpdateInfoRequest) String() string {
 func (*UpdateInfoRequest) ProtoMessage() {}
 
 func (x *UpdateInfoRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_proto_v1_message_user_proto_msgTypes[4]
+	mi := &file_proto_v1_message_user_proto_msgTypes[6]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -326,7 +405,7 @@ func (x *UpdateInfoRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use UpdateInfoRequest.ProtoReflect.Descriptor instead.
 func (*UpdateInfoRequest) Descriptor() ([]byte, []int) {
-	return file_proto_v1_message_user_proto_rawDescGZIP(), []int{4}
+	return file_proto_v1_message_user_proto_rawDescGZIP(), []int{6}
 }
 
 func (x *UpdateInfoRequest) GetFile() []byte {
@@ -348,7 +427,7 @@ type UpdateInfoResponse struct {
 func (x *UpdateInfoResponse) Reset() {
 	*x = UpdateInfoResponse{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_v1_message_user_proto_msgTypes[5]
+		mi := &file_proto_v1_message_user_proto_msgTypes[7]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -361,7 +440,7 @@ func (x *UpdateInfoResponse) String() string {
 func (*UpdateInfoResponse) ProtoMessage() {}
 
 func (x *UpdateInfoResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_proto_v1_message_user_proto_msgTypes[5]
+	mi := &file_proto_v1_message_user_proto_msgTypes[7]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -374,7 +453,7 @@ func (x *UpdateInfoResponse) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use UpdateInfoResponse.ProtoReflect.Descriptor instead.
 func (*UpdateInfoResponse) Descriptor() ([]byte, []int) {
-	return file_proto_v1_message_user_proto_rawDescGZIP(), []int{5}
+	return file_proto_v1_message_user_proto_rawDescGZIP(), []int{7}
 }
 
 func (x *UpdateInfoResponse) GetAffectedRows() int64 {
@@ -407,7 +486,7 @@ type NotifyRequest struct {
 func (x *NotifyRequest) Reset() {
 	*x = NotifyRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_v1_message_user_proto_msgTypes[6]
+		mi := &file_proto_v1_message_user_proto_msgTypes[8]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -420,7 +499,7 @@ func (x *NotifyRequest) String() string {
 func (*NotifyRequest) ProtoMessage() {}
 
 func (x *NotifyRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_proto_v1_message_user_proto_msgTypes[6]
+	mi := &file_proto_v1_message_user_proto_msgTypes[8]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -433,7 +512,7 @@ func (x *NotifyRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use NotifyRequest.ProtoReflect.Descriptor instead.
 func (*NotifyRequest) Descriptor() ([]byte, []int) {
-	return file_proto_v1_message_user_proto_rawDescGZIP(), []int{6}
+	return file_proto_v1_message_user_proto_rawDescGZIP(), []int{8}
 }
 
 func (x *NotifyRequest) GetMsg() string {
@@ -491,7 +570,7 @@ type NotifyResponse struct {
 func (x *NotifyResponse) Reset() {
 	*x = NotifyResponse{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_v1_message_user_proto_msgTypes[7]
+		mi := &file_proto_v1_message_user_proto_msgTypes[9]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -504,7 +583,7 @@ func (x *NotifyResponse) String() string {
 func (*NotifyResponse) ProtoMessage() {}
 
 func (x *NotifyResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_proto_v1_message_user_proto_msgTypes[7]
+	mi := &file_proto_v1_message_user_proto_msgTypes[9]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -517,7 +596,7 @@ func (x *NotifyResponse) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use NotifyResponse.ProtoReflect.Descriptor instead.
 func (*NotifyResponse) Descriptor() ([]byte, []int) {
-	return file_proto_v1_message_user_proto_rawDescGZIP(), []int{7}
+	return file_proto_v1_message_user_proto_rawDescGZIP(), []int{9}
 }
 
 func (x *NotifyResponse) GetIsSend() bool {
@@ -560,27 +639,30 @@ var file_proto_v1_message_user_proto_rawDesc = []byte{
 	0x6c, 0x6f, 0x67, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b,
 	0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62,
 	0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x09, 0x6c, 0x6f,
-	0x67, 0x69, 0x6e, 0x65, 0x64, 0x41, 0x74, 0x22, 0x27, 0x0a, 0x11, 0x55, 0x70, 0x64, 0x61, 0x74,
-	0x65, 0x49, 0x6e, 0x66, 0x6f, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x12, 0x0a, 0x04,
-	0x66, 0x69, 0x6c, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x04, 0x66, 0x69, 0x6c, 0x65,
-	0x22, 0x56, 0x0a, 0x12, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x49, 0x6e, 0x66, 0x6f, 0x52, 0x65,
-	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x22, 0x0a, 0x0c, 0x61, 0x66, 0x66, 0x65, 0x63, 0x74,
-	0x65, 0x64, 0x52, 0x6f, 0x77, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x03, 0x52, 0x0c, 0x61, 0x66,
-	0x66, 0x65, 0x63, 0x74, 0x65, 0x64, 0x52, 0x6f, 0x77, 0x73, 0x12, 0x1c, 0x0a, 0x09, 0x69, 0x73,
-	0x53, 0x75, 0x63, 0x63, 0x65, 0x65, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x08, 0x52, 0x09, 0x69,
-	0x73, 0x53, 0x75, 0x63, 0x63, 0x65, 0x65, 0x64, 0x22, 0x6d, 0x0a, 0x0d, 0x4e, 0x6f, 0x74, 0x69,
-	0x66, 0x79, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x15, 0x0a, 0x03, 0x6d, 0x73, 0x67,
-	0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x48, 0x01, 0x52, 0x03, 0x6d, 0x73, 0x67, 0x88, 0x01, 0x01,
-	0x12, 0x16, 0x0a, 0x05, 0x65, 0x6d, 0x61, 0x69, 0x6c, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x48,
-	0x00, 0x52, 0x05, 0x65, 0x6d, 0x61, 0x69, 0x6c, 0x12, 0x16, 0x0a, 0x05, 0x70, 0x68, 0x6f, 0x6e,
-	0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x05, 0x70, 0x68, 0x6f, 0x6e, 0x65,
-	0x42, 0x0d, 0x0a, 0x0b, 0x6e, 0x6f, 0x74, 0x69, 0x66, 0x79, 0x53, 0x74, 0x79, 0x6c, 0x65, 0x42,
-	0x06, 0x0a, 0x04, 0x5f, 0x6d, 0x73, 0x67, 0x22, 0x28, 0x0a, 0x0e, 0x4e, 0x6f, 0x74, 0x69, 0x66,
-	0x79, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x16, 0x0a, 0x06, 0x69, 0x73, 0x53,
-	0x65, 0x6e, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x08, 0x52, 0x06, 0x69, 0x73, 0x53, 0x65, 0x6e,
-	0x64, 0x42, 0x1c, 0x5a, 0x1a, 0x68, 0x65, 0x6c, 0x6c, 0x6f, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x2f,
-	0x61, 0x70, 0x69, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x76, 0x31, 0x3b, 0x76, 0x31, 0x62,
-	0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x67, 0x69, 0x6e, 0x65, 0x64, 0x41, 0x74, 0x22, 0x0f, 0x0a, 0x0d, 0x4c, 0x6f, 0x67, 0x6f, 0x75,
+	0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0x10, 0x0a, 0x0e, 0x4c, 0x6f, 0x67, 0x6f,
+	0x75, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x27, 0x0a, 0x11, 0x55, 0x70,
+	0x64, 0x61, 0x74, 0x65, 0x49, 0x6e, 0x66, 0x6f, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12,
+	0x12, 0x0a, 0x04, 0x66, 0x69, 0x6c, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x04, 0x66,
+	0x69, 0x6c, 0x65, 0x22, 0x56, 0x0a, 0x12, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x49, 0x6e, 0x66,
+	0x6f, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x22, 0x0a, 0x0c, 0x61, 0x66, 0x66,
+	0x65, 0x63, 0x74, 0x65, 0x64, 0x52, 0x6f, 0x77, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x03, 0x52,
+	0x0c, 0x61, 0x66, 0x66, 0x65, 0x63, 0x74, 0x65, 0x64, 0x52, 0x6f, 0x77, 0x73, 0x12, 0x1c, 0x0a,
+	0x09, 0x69, 0x73, 0x53, 0x75, 0x63, 0x63, 0x65, 0x65, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x08,
+	0x52, 0x09, 0x69, 0x73, 0x53, 0x75, 0x63, 0x63, 0x65, 0x65, 0x64, 0x22, 0x6d, 0x0a, 0x0d, 0x4e,
+	0x6f, 0x74, 0x69, 0x66, 0x79, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x15, 0x0a, 0x03,
+	0x6d, 0x73, 0x67, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x48, 0x01, 0x52, 0x03, 0x6d, 0x73, 0x67,
+	0x88, 0x01, 0x01, 0x12, 0x16, 0x0a, 0x05, 0x65, 0x6d, 0x61, 0x69, 0x6c, 0x18, 0x02, 0x20, 0x01,
+	0x28, 0x09, 0x48, 0x00, 0x52, 0x05, 0x65, 0x6d, 0x61, 0x69, 0x6c, 0x12, 0x16, 0x0a, 0x05, 0x70,
+	0x68, 0x6f, 0x6e, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x05, 0x70, 0x68,
+	0x6f, 0x6e, 0x65, 0x42, 0x0d, 0x0a, 0x0b, 0x6e, 0x6f, 0x74, 0x69, 0x66, 0x79, 0x53, 0x74, 0x79,
+	0x6c, 0x65, 0x42, 0x06, 0x0a, 0x04, 0x5f, 0x6d, 0x73, 0x67, 0x22, 0x28, 0x0a, 0x0e, 0x4e, 0x6f,
+	0x74, 0x69, 0x66, 0x79, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x16, 0x0a, 0x06,
+	0x69, 0x73, 0x53, 0x65, 0x6e, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x08, 0x52, 0x06, 0x69, 0x73,
+	0x53, 0x65, 0x6e, 0x64, 0x42, 0x2f, 0x5a, 0x2d, 0x67, 0x69, 0x74, 0x2e, 0x78, 0x71, 0x35, 0x2e,
+	0x63, 0x6f, 0x6d, 0x2f, 0x67, 0x6f, 0x6c, 0x61, 0x6e, 0x67, 0x2f, 0x68, 0x65, 0x6c, 0x6c, 0x6f,
+	0x77, 0x6f, 0x72, 0x6c, 0x64, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f,
+	0x76, 0x31, 0x3b, 0x76, 0x31, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -596,27 +678,29 @@ func file_proto_v1_message_user_proto_rawDescGZIP() []byte {
 }
 
 var file_proto_v1_message_user_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
-var file_proto_v1_message_user_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
+var file_proto_v1_message_user_proto_msgTypes = make([]protoimpl.MessageInfo, 10)
 var file_proto_v1_message_user_proto_goTypes = []interface{}{
 	(UserInfo_Gender)(0),          // 0: helloworld.api.v1.UserInfo.Gender
 	(*UserRequest)(nil),           // 1: helloworld.api.v1.UserRequest
 	(*UserInfo)(nil),              // 2: helloworld.api.v1.UserInfo
 	(*LoginRequest)(nil),          // 3: helloworld.api.v1.LoginRequest
 	(*LoginResponse)(nil),         // 4: helloworld.api.v1.LoginResponse
-	(*UpdateInfoRequest)(nil),     // 5: helloworld.api.v1.UpdateInfoRequest
-	(*UpdateInfoResponse)(nil),    // 6: helloworld.api.v1.UpdateInfoResponse
-	(*NotifyRequest)(nil),         // 7: helloworld.api.v1.NotifyRequest
-	(*NotifyResponse)(nil),        // 8: helloworld.api.v1.NotifyResponse
-	(*timestamppb.Timestamp)(nil), // 9: google.protobuf.Timestamp
+	(*LogoutRequest)(nil),         // 5: helloworld.api.v1.LogoutRequest
+	(*LogoutResponse)(nil),        // 6: helloworld.api.v1.LogoutResponse
+	(*UpdateInfoRequest)(nil),     // 7: helloworld.api.v1.UpdateInfoRequest
+	(*UpdateInfoResponse)(nil),    // 8: helloworld.api.v1.UpdateInfoResponse
+	(*NotifyRequest)(nil),         // 9: helloworld.api.v1.NotifyRequest
+	(*NotifyResponse)(nil),        // 10: helloworld.api.v1.NotifyResponse
+	(*timestamppb.Timestamp)(nil), // 11: google.protobuf.Timestamp
 }
 var file_proto_v1_message_user_proto_depIdxs = []int32{
-	0, // 0: helloworld.api.v1.UserInfo.gender:type_name -> helloworld.api.v1.UserInfo.Gender
-	9, // 1: helloworld.api.v1.LoginResponse.logined_at:type_name -> google.protobuf.Timestamp
-	2, // [2:2] is the sub-list for method output_type
-	2, // [2:2] is the sub-list for method input_type
-	2, // [2:2] is the sub-list for extension type_name
-	2, // [2:2] is the sub-list for extension extendee
-	0, // [0:2] is the sub-list for field type_name
+	0,  // 0: helloworld.api.v1.UserInfo.gender:type_name -> helloworld.api.v1.UserInfo.Gender
+	11, // 1: helloworld.api.v1.LoginResponse.logined_at:type_name -> google.protobuf.Timestamp
+	2,  // [2:2] is the sub-list for method output_type
+	2,  // [2:2] is the sub-list for method input_type
+	2,  // [2:2] is the sub-list for extension type_name
+	2,  // [2:2] is the sub-list for extension extendee
+	0,  // [0:2] is the sub-list for field type_name
 }
 
 func init() { file_proto_v1_message_user_proto_init() }
@@ -674,7 +758,7 @@ func file_proto_v1_message_user_proto_init() {
 			}
 		}
 		file_proto_v1_message_user_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*UpdateInfoRequest); i {
+			switch v := v.(*LogoutRequest); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -686,7 +770,7 @@ func file_proto_v1_message_user_proto_init() {
 			}
 		}
 		file_proto_v1_message_user_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*UpdateInfoResponse); i {
+			switch v := v.(*LogoutResponse); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -698,7 +782,7 @@ func file_proto_v1_message_user_proto_init() {
 			}
 		}
 		file_proto_v1_message_user_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*NotifyRequest); i {
+			switch v := v.(*UpdateInfoRequest); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -710,6 +794,30 @@ func file_proto_v1_message_user_proto_init() {
 			}
 		}
 		file_proto_v1_message_user_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*UpdateInfoResponse); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_proto_v1_message_user_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*NotifyRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_proto_v1_message_user_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
 			switch v := v.(*NotifyResponse); i {
 			case 0:
 				return &v.state
@@ -723,7 +831,7 @@ func file_proto_v1_message_user_proto_init() {
 		}
 	}
 	file_proto_v1_message_user_proto_msgTypes[2].OneofWrappers = []interface{}{}
-	file_proto_v1_message_user_proto_msgTypes[6].OneofWrappers = []interface{}{
+	file_proto_v1_message_user_proto_msgTypes[8].OneofWrappers = []interface{}{
 		(*NotifyRequest_Email)(nil),
 		(*NotifyRequest_Phone)(nil),
 	}
@@ -733,7 +841,7 @@ func file_proto_v1_message_user_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_proto_v1_message_user_proto_rawDesc,
 			NumEnums:      1,
-			NumMessages:   8,
+			NumMessages:   10,
 			NumExtensions: 0,
 			NumServices:   0,
 		},
