@@ -81,18 +81,19 @@ type Session struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	RedisConnectType string   `protobuf:"bytes,1,opt,name=RedisConnectType,proto3" json:"RedisConnectType,omitempty"`
-	Host             string   `protobuf:"bytes,2,opt,name=Host,proto3" json:"Host,omitempty"`
-	Port             int32    `protobuf:"varint,3,opt,name=Port,proto3" json:"Port,omitempty"`
-	Password         string   `protobuf:"bytes,4,opt,name=Password,proto3" json:"Password,omitempty"`
-	DbNumber         int32    `protobuf:"varint,5,opt,name=DbNumber,proto3" json:"DbNumber,omitempty"`
-	MaxIdle          int32    `protobuf:"varint,6,opt,name=MaxIdle,proto3" json:"MaxIdle,omitempty"`
-	Secret           string   `protobuf:"bytes,7,opt,name=Secret,proto3" json:"Secret,omitempty"`
-	SessionName      []string `protobuf:"bytes,8,rep,name=SessionName,proto3" json:"SessionName,omitempty"`
-	Path             string   `protobuf:"bytes,9,opt,name=Path,proto3" json:"Path,omitempty"`
-	MaxAge           int32    `protobuf:"varint,10,opt,name=MaxAge,proto3" json:"MaxAge,omitempty"`
-	HTTPOnly         bool     `protobuf:"varint,11,opt,name=HTTPOnly,proto3" json:"HTTPOnly,omitempty"`
-	SameSite         int32    `protobuf:"varint,12,opt,name=SameSite,proto3" json:"SameSite,omitempty"`
+	RedisConnectType string   `protobuf:"bytes,1,opt,name=redisConnectType,proto3" json:"redisConnectType,omitempty"`
+	Host             string   `protobuf:"bytes,2,opt,name=host,proto3" json:"host,omitempty"`
+	Port             int32    `protobuf:"varint,3,opt,name=port,proto3" json:"port,omitempty"`
+	Password         string   `protobuf:"bytes,4,opt,name=password,proto3" json:"password,omitempty"`
+	DbNumber         int32    `protobuf:"varint,5,opt,name=dbNumber,proto3" json:"dbNumber,omitempty"`
+	MaxIdle          int32    `protobuf:"varint,6,opt,name=maxIdle,proto3" json:"maxIdle,omitempty"`
+	Secret           []string `protobuf:"bytes,7,rep,name=secret,proto3" json:"secret,omitempty"`
+	// 支持配置多个sessionName key是代指标签 value是session对应的名称
+	SessionName map[string]string `protobuf:"bytes,8,rep,name=sessionName,proto3" json:"sessionName,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	Path        string            `protobuf:"bytes,9,opt,name=path,proto3" json:"path,omitempty"`
+	MaxAge      int32             `protobuf:"varint,10,opt,name=maxAge,proto3" json:"maxAge,omitempty"`
+	HttpOnly    bool              `protobuf:"varint,11,opt,name=httpOnly,proto3" json:"httpOnly,omitempty"`
+	SameSite    int32             `protobuf:"varint,12,opt,name=sameSite,proto3" json:"sameSite,omitempty"`
 }
 
 func (x *Session) Reset() {
@@ -169,14 +170,14 @@ func (x *Session) GetMaxIdle() int32 {
 	return 0
 }
 
-func (x *Session) GetSecret() string {
+func (x *Session) GetSecret() []string {
 	if x != nil {
 		return x.Secret
 	}
-	return ""
+	return nil
 }
 
-func (x *Session) GetSessionName() []string {
+func (x *Session) GetSessionName() map[string]string {
 	if x != nil {
 		return x.SessionName
 	}
@@ -197,9 +198,9 @@ func (x *Session) GetMaxAge() int32 {
 	return 0
 }
 
-func (x *Session) GetHTTPOnly() bool {
+func (x *Session) GetHttpOnly() bool {
 	if x != nil {
-		return x.HTTPOnly
+		return x.HttpOnly
 	}
 	return false
 }
@@ -216,8 +217,8 @@ type APPConfig struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	RunMode string  `protobuf:"bytes,1,opt,name=RunMode,proto3" json:"RunMode,omitempty"` // 运行模式
-	Switch  *Switch `protobuf:"bytes,2,opt,name=Switch,proto3" json:"Switch,omitempty"`   // 开关
+	RunMode string  `protobuf:"bytes,1,opt,name=runMode,proto3" json:"runMode,omitempty"` // 运行模式
+	Switch  *Switch `protobuf:"bytes,2,opt,name=switch,proto3" json:"switch,omitempty"`   // 开关
 }
 
 func (x *APPConfig) Reset() {
@@ -271,10 +272,10 @@ type Switch struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	CloseSendSMS    bool `protobuf:"varint,1,opt,name=CloseSendSMS,proto3" json:"CloseSendSMS,omitempty"`       //关闭发送短信
-	CloseSendEmail  bool `protobuf:"varint,2,opt,name=CloseSendEmail,proto3" json:"CloseSendEmail,omitempty"`   //关闭发送邮件
-	CloseValidSMS   bool `protobuf:"varint,3,opt,name=CloseValidSMS,proto3" json:"CloseValidSMS,omitempty"`     //关闭验证短信
-	CloseValidEmail bool `protobuf:"varint,4,opt,name=CloseValidEmail,proto3" json:"CloseValidEmail,omitempty"` //关闭验证邮件
+	CloseSendSMS    bool `protobuf:"varint,1,opt,name=CloseSendSMS,proto3" json:"CloseSendSMS,omitempty"`       // 关闭发送短信
+	CloseSendEmail  bool `protobuf:"varint,2,opt,name=CloseSendEmail,proto3" json:"CloseSendEmail,omitempty"`   // 关闭发送邮件
+	CloseValidSMS   bool `protobuf:"varint,3,opt,name=CloseValidSMS,proto3" json:"CloseValidSMS,omitempty"`     // 关闭验证短信
+	CloseValidEmail bool `protobuf:"varint,4,opt,name=CloseValidEmail,proto3" json:"CloseValidEmail,omitempty"` // 关闭验证邮件
 }
 
 func (x *Switch) Reset() {
@@ -337,6 +338,81 @@ func (x *Switch) GetCloseValidEmail() bool {
 	return false
 }
 
+type Hours struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	// @gotags: toml:"serverName"
+	ServerName string `protobuf:"bytes,1,opt,name=serverName,proto3" json:"serverName,omitempty" toml:"serverName"`
+	// @gotags: toml:"endpoints"
+	Endpoints []string `protobuf:"bytes,2,rep,name=endpoints,proto3" json:"endpoints,omitempty" toml:"endpoints"`
+	// @gotags: toml:"ports"
+	Ports map[string]string `protobuf:"bytes,3,rep,name=ports,proto3" json:"ports,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3" toml:"ports"`
+	// @gotags: toml:"disablePProf"
+	DisablePProf bool `protobuf:"varint,4,opt,name=disablePProf,proto3" json:"disablePProf,omitempty" toml:"disablePProf"`
+}
+
+func (x *Hours) Reset() {
+	*x = Hours{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_conf_conf_proto_msgTypes[4]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *Hours) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*Hours) ProtoMessage() {}
+
+func (x *Hours) ProtoReflect() protoreflect.Message {
+	mi := &file_conf_conf_proto_msgTypes[4]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use Hours.ProtoReflect.Descriptor instead.
+func (*Hours) Descriptor() ([]byte, []int) {
+	return file_conf_conf_proto_rawDescGZIP(), []int{4}
+}
+
+func (x *Hours) GetServerName() string {
+	if x != nil {
+		return x.ServerName
+	}
+	return ""
+}
+
+func (x *Hours) GetEndpoints() []string {
+	if x != nil {
+		return x.Endpoints
+	}
+	return nil
+}
+
+func (x *Hours) GetPorts() map[string]string {
+	if x != nil {
+		return x.Ports
+	}
+	return nil
+}
+
+func (x *Hours) GetDisablePProf() bool {
+	if x != nil {
+		return x.DisablePProf
+	}
+	return false
+}
+
 var File_conf_conf_proto protoreflect.FileDescriptor
 
 var file_conf_conf_proto_rawDesc = []byte{
@@ -348,46 +424,67 @@ var file_conf_conf_proto_rawDesc = []byte{
 	0x01, 0x28, 0x09, 0x52, 0x04, 0x61, 0x64, 0x64, 0x72, 0x12, 0x33, 0x0a, 0x07, 0x74, 0x69, 0x6d,
 	0x65, 0x6f, 0x75, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x19, 0x2e, 0x67, 0x6f, 0x6f,
 	0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x44, 0x75, 0x72,
-	0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x07, 0x74, 0x69, 0x6d, 0x65, 0x6f, 0x75, 0x74, 0x22, 0xcd,
-	0x02, 0x0a, 0x07, 0x53, 0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x2a, 0x0a, 0x10, 0x52, 0x65,
+	0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x07, 0x74, 0x69, 0x6d, 0x65, 0x6f, 0x75, 0x74, 0x22, 0xb1,
+	0x03, 0x0a, 0x07, 0x53, 0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x2a, 0x0a, 0x10, 0x72, 0x65,
 	0x64, 0x69, 0x73, 0x43, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x54, 0x79, 0x70, 0x65, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x10, 0x52, 0x65, 0x64, 0x69, 0x73, 0x43, 0x6f, 0x6e, 0x6e, 0x65,
-	0x63, 0x74, 0x54, 0x79, 0x70, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x48, 0x6f, 0x73, 0x74, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x48, 0x6f, 0x73, 0x74, 0x12, 0x12, 0x0a, 0x04, 0x50, 0x6f,
-	0x72, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x05, 0x52, 0x04, 0x50, 0x6f, 0x72, 0x74, 0x12, 0x1a,
-	0x0a, 0x08, 0x50, 0x61, 0x73, 0x73, 0x77, 0x6f, 0x72, 0x64, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x08, 0x50, 0x61, 0x73, 0x73, 0x77, 0x6f, 0x72, 0x64, 0x12, 0x1a, 0x0a, 0x08, 0x44, 0x62,
-	0x4e, 0x75, 0x6d, 0x62, 0x65, 0x72, 0x18, 0x05, 0x20, 0x01, 0x28, 0x05, 0x52, 0x08, 0x44, 0x62,
-	0x4e, 0x75, 0x6d, 0x62, 0x65, 0x72, 0x12, 0x18, 0x0a, 0x07, 0x4d, 0x61, 0x78, 0x49, 0x64, 0x6c,
-	0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x05, 0x52, 0x07, 0x4d, 0x61, 0x78, 0x49, 0x64, 0x6c, 0x65,
-	0x12, 0x16, 0x0a, 0x06, 0x53, 0x65, 0x63, 0x72, 0x65, 0x74, 0x18, 0x07, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x06, 0x53, 0x65, 0x63, 0x72, 0x65, 0x74, 0x12, 0x20, 0x0a, 0x0b, 0x53, 0x65, 0x73, 0x73,
-	0x69, 0x6f, 0x6e, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x08, 0x20, 0x03, 0x28, 0x09, 0x52, 0x0b, 0x53,
-	0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x50, 0x61,
-	0x74, 0x68, 0x18, 0x09, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x50, 0x61, 0x74, 0x68, 0x12, 0x16,
-	0x0a, 0x06, 0x4d, 0x61, 0x78, 0x41, 0x67, 0x65, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x05, 0x52, 0x06,
-	0x4d, 0x61, 0x78, 0x41, 0x67, 0x65, 0x12, 0x1a, 0x0a, 0x08, 0x48, 0x54, 0x54, 0x50, 0x4f, 0x6e,
-	0x6c, 0x79, 0x18, 0x0b, 0x20, 0x01, 0x28, 0x08, 0x52, 0x08, 0x48, 0x54, 0x54, 0x50, 0x4f, 0x6e,
-	0x6c, 0x79, 0x12, 0x1a, 0x0a, 0x08, 0x53, 0x61, 0x6d, 0x65, 0x53, 0x69, 0x74, 0x65, 0x18, 0x0c,
-	0x20, 0x01, 0x28, 0x05, 0x52, 0x08, 0x53, 0x61, 0x6d, 0x65, 0x53, 0x69, 0x74, 0x65, 0x22, 0x4f,
-	0x0a, 0x09, 0x41, 0x50, 0x50, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x12, 0x18, 0x0a, 0x07, 0x52,
-	0x75, 0x6e, 0x4d, 0x6f, 0x64, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x52, 0x75,
-	0x6e, 0x4d, 0x6f, 0x64, 0x65, 0x12, 0x28, 0x0a, 0x06, 0x53, 0x77, 0x69, 0x74, 0x63, 0x68, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x63, 0x6f, 0x6e, 0x66, 0x2e, 0x61, 0x70, 0x69,
-	0x2e, 0x53, 0x77, 0x69, 0x74, 0x63, 0x68, 0x52, 0x06, 0x53, 0x77, 0x69, 0x74, 0x63, 0x68, 0x22,
-	0xa4, 0x01, 0x0a, 0x06, 0x53, 0x77, 0x69, 0x74, 0x63, 0x68, 0x12, 0x22, 0x0a, 0x0c, 0x43, 0x6c,
-	0x6f, 0x73, 0x65, 0x53, 0x65, 0x6e, 0x64, 0x53, 0x4d, 0x53, 0x18, 0x01, 0x20, 0x01, 0x28, 0x08,
-	0x52, 0x0c, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x53, 0x65, 0x6e, 0x64, 0x53, 0x4d, 0x53, 0x12, 0x26,
-	0x0a, 0x0e, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x53, 0x65, 0x6e, 0x64, 0x45, 0x6d, 0x61, 0x69, 0x6c,
-	0x18, 0x02, 0x20, 0x01, 0x28, 0x08, 0x52, 0x0e, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x53, 0x65, 0x6e,
-	0x64, 0x45, 0x6d, 0x61, 0x69, 0x6c, 0x12, 0x24, 0x0a, 0x0d, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x56,
-	0x61, 0x6c, 0x69, 0x64, 0x53, 0x4d, 0x53, 0x18, 0x03, 0x20, 0x01, 0x28, 0x08, 0x52, 0x0d, 0x43,
-	0x6c, 0x6f, 0x73, 0x65, 0x56, 0x61, 0x6c, 0x69, 0x64, 0x53, 0x4d, 0x53, 0x12, 0x28, 0x0a, 0x0f,
-	0x43, 0x6c, 0x6f, 0x73, 0x65, 0x56, 0x61, 0x6c, 0x69, 0x64, 0x45, 0x6d, 0x61, 0x69, 0x6c, 0x18,
-	0x04, 0x20, 0x01, 0x28, 0x08, 0x52, 0x0f, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x56, 0x61, 0x6c, 0x69,
-	0x64, 0x45, 0x6d, 0x61, 0x69, 0x6c, 0x42, 0x1f, 0x5a, 0x1d, 0x68, 0x65, 0x6c, 0x6c, 0x6f, 0x77,
-	0x6f, 0x72, 0x6c, 0x64, 0x2f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x2f, 0x63, 0x6f,
-	0x6e, 0x66, 0x3b, 0x63, 0x6f, 0x6e, 0x66, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x10, 0x72, 0x65, 0x64, 0x69, 0x73, 0x43, 0x6f, 0x6e, 0x6e, 0x65,
+	0x63, 0x74, 0x54, 0x79, 0x70, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x68, 0x6f, 0x73, 0x74, 0x18, 0x02,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x68, 0x6f, 0x73, 0x74, 0x12, 0x12, 0x0a, 0x04, 0x70, 0x6f,
+	0x72, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x05, 0x52, 0x04, 0x70, 0x6f, 0x72, 0x74, 0x12, 0x1a,
+	0x0a, 0x08, 0x70, 0x61, 0x73, 0x73, 0x77, 0x6f, 0x72, 0x64, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09,
+	0x52, 0x08, 0x70, 0x61, 0x73, 0x73, 0x77, 0x6f, 0x72, 0x64, 0x12, 0x1a, 0x0a, 0x08, 0x64, 0x62,
+	0x4e, 0x75, 0x6d, 0x62, 0x65, 0x72, 0x18, 0x05, 0x20, 0x01, 0x28, 0x05, 0x52, 0x08, 0x64, 0x62,
+	0x4e, 0x75, 0x6d, 0x62, 0x65, 0x72, 0x12, 0x18, 0x0a, 0x07, 0x6d, 0x61, 0x78, 0x49, 0x64, 0x6c,
+	0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x05, 0x52, 0x07, 0x6d, 0x61, 0x78, 0x49, 0x64, 0x6c, 0x65,
+	0x12, 0x16, 0x0a, 0x06, 0x73, 0x65, 0x63, 0x72, 0x65, 0x74, 0x18, 0x07, 0x20, 0x03, 0x28, 0x09,
+	0x52, 0x06, 0x73, 0x65, 0x63, 0x72, 0x65, 0x74, 0x12, 0x44, 0x0a, 0x0b, 0x73, 0x65, 0x73, 0x73,
+	0x69, 0x6f, 0x6e, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x08, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x22, 0x2e,
+	0x63, 0x6f, 0x6e, 0x66, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e,
+	0x2e, 0x53, 0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x4e, 0x61, 0x6d, 0x65, 0x45, 0x6e, 0x74, 0x72,
+	0x79, 0x52, 0x0b, 0x73, 0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x12,
+	0x0a, 0x04, 0x70, 0x61, 0x74, 0x68, 0x18, 0x09, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x70, 0x61,
+	0x74, 0x68, 0x12, 0x16, 0x0a, 0x06, 0x6d, 0x61, 0x78, 0x41, 0x67, 0x65, 0x18, 0x0a, 0x20, 0x01,
+	0x28, 0x05, 0x52, 0x06, 0x6d, 0x61, 0x78, 0x41, 0x67, 0x65, 0x12, 0x1a, 0x0a, 0x08, 0x68, 0x74,
+	0x74, 0x70, 0x4f, 0x6e, 0x6c, 0x79, 0x18, 0x0b, 0x20, 0x01, 0x28, 0x08, 0x52, 0x08, 0x68, 0x74,
+	0x74, 0x70, 0x4f, 0x6e, 0x6c, 0x79, 0x12, 0x1a, 0x0a, 0x08, 0x73, 0x61, 0x6d, 0x65, 0x53, 0x69,
+	0x74, 0x65, 0x18, 0x0c, 0x20, 0x01, 0x28, 0x05, 0x52, 0x08, 0x73, 0x61, 0x6d, 0x65, 0x53, 0x69,
+	0x74, 0x65, 0x1a, 0x3e, 0x0a, 0x10, 0x53, 0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x4e, 0x61, 0x6d,
+	0x65, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x03, 0x6b, 0x65, 0x79, 0x12, 0x14, 0x0a, 0x05, 0x76, 0x61, 0x6c, 0x75,
+	0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a, 0x02,
+	0x38, 0x01, 0x22, 0x4f, 0x0a, 0x09, 0x41, 0x50, 0x50, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x12,
+	0x18, 0x0a, 0x07, 0x72, 0x75, 0x6e, 0x4d, 0x6f, 0x64, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09,
+	0x52, 0x07, 0x72, 0x75, 0x6e, 0x4d, 0x6f, 0x64, 0x65, 0x12, 0x28, 0x0a, 0x06, 0x73, 0x77, 0x69,
+	0x74, 0x63, 0x68, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x63, 0x6f, 0x6e, 0x66,
+	0x2e, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x77, 0x69, 0x74, 0x63, 0x68, 0x52, 0x06, 0x73, 0x77, 0x69,
+	0x74, 0x63, 0x68, 0x22, 0xa4, 0x01, 0x0a, 0x06, 0x53, 0x77, 0x69, 0x74, 0x63, 0x68, 0x12, 0x22,
+	0x0a, 0x0c, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x53, 0x65, 0x6e, 0x64, 0x53, 0x4d, 0x53, 0x18, 0x01,
+	0x20, 0x01, 0x28, 0x08, 0x52, 0x0c, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x53, 0x65, 0x6e, 0x64, 0x53,
+	0x4d, 0x53, 0x12, 0x26, 0x0a, 0x0e, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x53, 0x65, 0x6e, 0x64, 0x45,
+	0x6d, 0x61, 0x69, 0x6c, 0x18, 0x02, 0x20, 0x01, 0x28, 0x08, 0x52, 0x0e, 0x43, 0x6c, 0x6f, 0x73,
+	0x65, 0x53, 0x65, 0x6e, 0x64, 0x45, 0x6d, 0x61, 0x69, 0x6c, 0x12, 0x24, 0x0a, 0x0d, 0x43, 0x6c,
+	0x6f, 0x73, 0x65, 0x56, 0x61, 0x6c, 0x69, 0x64, 0x53, 0x4d, 0x53, 0x18, 0x03, 0x20, 0x01, 0x28,
+	0x08, 0x52, 0x0d, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x56, 0x61, 0x6c, 0x69, 0x64, 0x53, 0x4d, 0x53,
+	0x12, 0x28, 0x0a, 0x0f, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x56, 0x61, 0x6c, 0x69, 0x64, 0x45, 0x6d,
+	0x61, 0x69, 0x6c, 0x18, 0x04, 0x20, 0x01, 0x28, 0x08, 0x52, 0x0f, 0x43, 0x6c, 0x6f, 0x73, 0x65,
+	0x56, 0x61, 0x6c, 0x69, 0x64, 0x45, 0x6d, 0x61, 0x69, 0x6c, 0x22, 0xd5, 0x01, 0x0a, 0x05, 0x48,
+	0x6f, 0x75, 0x72, 0x73, 0x12, 0x1e, 0x0a, 0x0a, 0x73, 0x65, 0x72, 0x76, 0x65, 0x72, 0x4e, 0x61,
+	0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x73, 0x65, 0x72, 0x76, 0x65, 0x72,
+	0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1c, 0x0a, 0x09, 0x65, 0x6e, 0x64, 0x70, 0x6f, 0x69, 0x6e, 0x74,
+	0x73, 0x18, 0x02, 0x20, 0x03, 0x28, 0x09, 0x52, 0x09, 0x65, 0x6e, 0x64, 0x70, 0x6f, 0x69, 0x6e,
+	0x74, 0x73, 0x12, 0x30, 0x0a, 0x05, 0x70, 0x6f, 0x72, 0x74, 0x73, 0x18, 0x03, 0x20, 0x03, 0x28,
+	0x0b, 0x32, 0x1a, 0x2e, 0x63, 0x6f, 0x6e, 0x66, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x48, 0x6f, 0x75,
+	0x72, 0x73, 0x2e, 0x50, 0x6f, 0x72, 0x74, 0x73, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x05, 0x70,
+	0x6f, 0x72, 0x74, 0x73, 0x12, 0x22, 0x0a, 0x0c, 0x64, 0x69, 0x73, 0x61, 0x62, 0x6c, 0x65, 0x50,
+	0x50, 0x72, 0x6f, 0x66, 0x18, 0x04, 0x20, 0x01, 0x28, 0x08, 0x52, 0x0c, 0x64, 0x69, 0x73, 0x61,
+	0x62, 0x6c, 0x65, 0x50, 0x50, 0x72, 0x6f, 0x66, 0x1a, 0x38, 0x0a, 0x0a, 0x50, 0x6f, 0x72, 0x74,
+	0x73, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x03, 0x6b, 0x65, 0x79, 0x12, 0x14, 0x0a, 0x05, 0x76, 0x61, 0x6c, 0x75,
+	0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a, 0x02,
+	0x38, 0x01, 0x42, 0x32, 0x5a, 0x30, 0x67, 0x69, 0x74, 0x2e, 0x78, 0x71, 0x35, 0x2e, 0x63, 0x6f,
+	0x6d, 0x2f, 0x67, 0x6f, 0x6c, 0x61, 0x6e, 0x67, 0x2f, 0x68, 0x65, 0x6c, 0x6c, 0x6f, 0x77, 0x6f,
+	0x72, 0x6c, 0x64, 0x2f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x2f, 0x63, 0x6f, 0x6e,
+	0x66, 0x3b, 0x63, 0x6f, 0x6e, 0x66, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -402,22 +499,27 @@ func file_conf_conf_proto_rawDescGZIP() []byte {
 	return file_conf_conf_proto_rawDescData
 }
 
-var file_conf_conf_proto_msgTypes = make([]protoimpl.MessageInfo, 4)
+var file_conf_conf_proto_msgTypes = make([]protoimpl.MessageInfo, 7)
 var file_conf_conf_proto_goTypes = []interface{}{
 	(*Server)(nil),              // 0: conf.api.Server
 	(*Session)(nil),             // 1: conf.api.Session
 	(*APPConfig)(nil),           // 2: conf.api.APPConfig
 	(*Switch)(nil),              // 3: conf.api.Switch
-	(*durationpb.Duration)(nil), // 4: google.protobuf.Duration
+	(*Hours)(nil),               // 4: conf.api.Hours
+	nil,                         // 5: conf.api.Session.SessionNameEntry
+	nil,                         // 6: conf.api.Hours.PortsEntry
+	(*durationpb.Duration)(nil), // 7: google.protobuf.Duration
 }
 var file_conf_conf_proto_depIdxs = []int32{
-	4, // 0: conf.api.Server.timeout:type_name -> google.protobuf.Duration
-	3, // 1: conf.api.APPConfig.Switch:type_name -> conf.api.Switch
-	2, // [2:2] is the sub-list for method output_type
-	2, // [2:2] is the sub-list for method input_type
-	2, // [2:2] is the sub-list for extension type_name
-	2, // [2:2] is the sub-list for extension extendee
-	0, // [0:2] is the sub-list for field type_name
+	7, // 0: conf.api.Server.timeout:type_name -> google.protobuf.Duration
+	5, // 1: conf.api.Session.sessionName:type_name -> conf.api.Session.SessionNameEntry
+	3, // 2: conf.api.APPConfig.switch:type_name -> conf.api.Switch
+	6, // 3: conf.api.Hours.ports:type_name -> conf.api.Hours.PortsEntry
+	4, // [4:4] is the sub-list for method output_type
+	4, // [4:4] is the sub-list for method input_type
+	4, // [4:4] is the sub-list for extension type_name
+	4, // [4:4] is the sub-list for extension extendee
+	0, // [0:4] is the sub-list for field type_name
 }
 
 func init() { file_conf_conf_proto_init() }
@@ -474,6 +576,18 @@ func file_conf_conf_proto_init() {
 				return nil
 			}
 		}
+		file_conf_conf_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*Hours); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
@@ -481,7 +595,7 @@ func file_conf_conf_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_conf_conf_proto_rawDesc,
 			NumEnums:      0,
-			NumMessages:   4,
+			NumMessages:   7,
 			NumExtensions: 0,
 			NumServices:   0,
 		},
